cleanup, reorganize
