+++ front/lib/models/assistant/agent.ts
@@ -1,8 +1,4 @@
-import type {
-  AgentUserListStatus,
-  DustAppRunConfigurationType,
-  TimeframeUnit,
-} from "@dust-tt/types";
+import type { AgentUserListStatus } from "@dust-tt/types";
 import type {
   AgentConfigurationScope,
   AgentStatus,
@@ -107,22 +103,9 @@ export class AgentConfiguration extends Model<
   declare generationConfigurationId: ForeignKey<
     AgentGenerationConfiguration["id"]
   > | null;
-  declare retrievalConfigurationId: ForeignKey<
-    AgentRetrievalConfiguration["id"]
-  > | null;
-  declare dustAppRunConfigurationId: ForeignKey<
-    AgentDustAppRunConfiguration["id"]
-  > | null;
-
-  declare tablesQueryConfigurationId: ForeignKey<
-    AgentTablesQueryConfiguration["id"]
-  > | null;
 
   declare author: NonAttribute<User>;
   declare generationConfiguration: NonAttribute<AgentGenerationConfiguration>;
-  declare retrievalConfiguration: NonAttribute<AgentRetrievalConfiguration>;
-  declare dustAppRunConfiguration: NonAttribute<DustAppRunConfigurationType>;
-  declare tablesQueryConfiguration: NonAttribute<AgentTablesQueryConfiguration>;
 }
 AgentConfiguration.init(
   {
@@ -358,286 +341,4 @@ AgentUserRelation.belongsTo(User, {
 });
 AgentUserRelation.belongsTo(Workspace, {
   foreignKey: { allowNull: false },
-});
-
-// TODO(@fontanierh) TO BE MOVED TO THE tables_query.ts file -- inlined during multi actions migration
-// to avoid circular dependencies
-
-export class AgentTablesQueryConfiguration extends Model<
-  InferAttributes<AgentTablesQueryConfiguration>,
-  InferCreationAttributes<AgentTablesQueryConfiguration>
-> {
-  declare id: CreationOptional<number>;
-  declare createdAt: CreationOptional<Date>;
-  declare updatedAt: CreationOptional<Date>;
-
-  declare agentConfigurationId: ForeignKey<AgentConfiguration["id"]>;
-
-  declare sId: string;
-}
-
-AgentTablesQueryConfiguration.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    sId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  },
-  {
-    modelName: "agent_tables_query_configuration",
-    indexes: [
-      {
-        unique: true,
-        fields: ["sId"],
-        name: "agent_tables_query_configuration_s_id",
-      },
-    ],
-    sequelize: frontSequelize,
-  }
-);
-
-// DEPERECATED -- AgentConfig -> TablesQueryConfig (1:1)
-AgentTablesQueryConfiguration.hasOne(AgentConfiguration, {
-  as: "tablesQueryConfiguration",
-  foreignKey: { name: "tablesQueryConfigurationId", allowNull: true }, // null = no Tables action set for this Agent
-});
-AgentConfiguration.belongsTo(AgentTablesQueryConfiguration, {
-  as: "tablesQueryConfiguration",
-  foreignKey: { name: "tablesQueryConfigurationId", allowNull: true }, // null = no Tables action set for this Agent
-});
-
-// NEW -- AgentConfig -> TablesQueryConfig (1:N)
-AgentConfiguration.hasMany(AgentTablesQueryConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
-});
-AgentTablesQueryConfiguration.belongsTo(AgentConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
-});
-
-// TODO(@fontanierh) TO BE MOVED TO THE retrieval.ts file -- inlined during multi actions migration
-// to avoid circular dependencies
-
-export class AgentRetrievalConfiguration extends Model<
-  InferAttributes<AgentRetrievalConfiguration>,
-  InferCreationAttributes<AgentRetrievalConfiguration>
-> {
-  declare id: CreationOptional<number>;
-  declare createdAt: CreationOptional<Date>;
-  declare updatedAt: CreationOptional<Date>;
-
-  declare agentConfigurationId: ForeignKey<AgentConfiguration["id"]>;
-
-  declare sId: string;
-
-  declare query: "auto" | "none" | "templated";
-  declare queryTemplate: string | null;
-  declare relativeTimeFrame: "auto" | "none" | "custom";
-  declare relativeTimeFrameDuration: number | null;
-  declare relativeTimeFrameUnit: TimeframeUnit | null;
-  declare topK: number | null;
-  declare topKMode: "auto" | "custom";
-}
-
-AgentRetrievalConfiguration.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    sId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    query: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      defaultValue: "auto",
-    },
-    queryTemplate: {
-      type: DataTypes.TEXT,
-      allowNull: true,
-    },
-    relativeTimeFrame: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      defaultValue: "auto",
-    },
-    relativeTimeFrameDuration: {
-      type: DataTypes.INTEGER,
-      allowNull: true,
-    },
-    relativeTimeFrameUnit: {
-      type: DataTypes.STRING,
-      allowNull: true,
-    },
-    topK: {
-      type: DataTypes.INTEGER,
-      allowNull: true,
-    },
-    topKMode: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      defaultValue: "auto",
-    },
-  },
-  {
-    modelName: "agent_retrieval_configuration",
-    sequelize: frontSequelize,
-    hooks: {
-      beforeValidate: (retrieval: AgentRetrievalConfiguration) => {
-        // Validation for templated Query
-        if (retrieval.query == "templated") {
-          if (!retrieval.queryTemplate) {
-            throw new Error("Must set a template for templated query");
-          }
-        } else if (retrieval.queryTemplate) {
-          throw new Error("Can't set a template without templated query");
-        }
-
-        // Validation for Timeframe
-        if (retrieval.relativeTimeFrame == "custom") {
-          if (
-            !retrieval.relativeTimeFrameDuration ||
-            !retrieval.relativeTimeFrameUnit
-          ) {
-            throw new Error(
-              "Custom relative time frame must have a duration and unit set"
-            );
-          }
-        }
-
-        // Validation for TopK
-        if (retrieval.topKMode == "custom") {
-          if (!retrieval.topK) {
-            throw new Error("topK must be set when topKMode is 'custom'");
-          }
-        } else if (retrieval.topK) {
-          throw new Error("topK must be null when topKMode is not 'custom'");
-        }
-      },
-    },
-  }
-);
-
-// DEPERECATED -- AgentConfig -> RetrievalConfig (1:1)
-AgentRetrievalConfiguration.hasOne(AgentConfiguration, {
-  as: "retrievalConfiguration",
-  foreignKey: { name: "retrievalConfigurationId", allowNull: true }, // null = no retrieval action set for this Agent
-});
-AgentConfiguration.belongsTo(AgentRetrievalConfiguration, {
-  as: "retrievalConfiguration",
-  foreignKey: { name: "retrievalConfigurationId", allowNull: true }, // null = no retrieval action set for this Agent
-});
-
-// NEW -- AgentConfig -> RetrievalConfig (1:N)
-AgentConfiguration.hasMany(AgentRetrievalConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
-});
-AgentRetrievalConfiguration.belongsTo(AgentConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
-});
-
-// TODO(@fontanierh) TO BE MOVED TO THE dust_app_run.ts file -- inlined during multi actions migration
-// to avoid circular dependencies
-
-export class AgentDustAppRunConfiguration extends Model<
-  InferAttributes<AgentDustAppRunConfiguration>,
-  InferCreationAttributes<AgentDustAppRunConfiguration>
-> {
-  declare id: CreationOptional<number>;
-  declare createdAt: CreationOptional<Date>;
-  declare updatedAt: CreationOptional<Date>;
-
-  declare agentConfigurationId: ForeignKey<AgentConfiguration["id"]>;
-
-  declare sId: string;
-
-  declare appWorkspaceId: string;
-  declare appId: string;
-}
-
-AgentDustAppRunConfiguration.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    sId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    appWorkspaceId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    appId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  },
-  {
-    modelName: "agent_dust_app_run_configuration",
-    indexes: [
-      {
-        unique: true,
-        fields: ["sId"],
-      },
-    ],
-    sequelize: frontSequelize,
-  }
-);
-
-// DEPERECATED -- AgentConfig -> DustAppRunConfig (1:1)
-AgentDustAppRunConfiguration.hasOne(AgentConfiguration, {
-  as: "dustAppRunConfiguration",
-  foreignKey: { name: "dustAppRunConfigurationId", allowNull: true }, // null = no DustAppRun action set for this Agent
-});
-AgentConfiguration.belongsTo(AgentDustAppRunConfiguration, {
-  as: "dustAppRunConfiguration",
-  foreignKey: { name: "dustAppRunConfigurationId", allowNull: true }, // null = no DustAppRun action set for this Agent
-});
-
-// NEW -- AgentConfig -> DustAppRunConfig (1:N)
-AgentConfiguration.hasMany(AgentDustAppRunConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
-});
-AgentDustAppRunConfiguration.belongsTo(AgentConfiguration, {
-  foreignKey: { name: "agentConfigurationId", allowNull: false },
 });
