+++ front/lib/models/index.ts
@@ -1,21 +1,23 @@
 import { App, Clone, Dataset, Provider, Run } from "@app/lib/models/apps";
-import { AgentDustAppRunAction } from "@app/lib/models/assistant/actions/dust_app_run";
+import {
+  AgentDustAppRunAction,
+  AgentDustAppRunConfiguration,
+} from "@app/lib/models/assistant/actions/dust_app_run";
 import {
   AgentDataSourceConfiguration,
   AgentRetrievalAction,
+  AgentRetrievalConfiguration,
   RetrievalDocument,
   RetrievalDocumentChunk,
 } from "@app/lib/models/assistant/actions/retrieval";
 import {
   AgentTablesQueryAction,
+  AgentTablesQueryConfiguration,
   AgentTablesQueryConfigurationTable,
 } from "@app/lib/models/assistant/actions/tables_query";
 import {
   AgentConfiguration,
-  AgentDustAppRunConfiguration,
   AgentGenerationConfiguration,
-  AgentRetrievalConfiguration,
-  AgentTablesQueryConfiguration,
   AgentUserRelation,
   GlobalAgentSettings,
 } from "@app/lib/models/assistant/agent";
